Release 3.0.0. Get rid of RSG.Promise and RestSharp, use typical classic approach.
