+++ lib/portable-net45+netcore45+wpa81+wp8/Voucherify.Client.dll
