+++ src/Voucherify.Client/Api/ApiResponse.cs
