+++ src/Voucherify.Client/Api/Customers.net45.cs
