+++ src/Voucherify.Client/Api/Redemptions.net45.cs
