+++ src/Voucherify.Client/Voucherify.Client.net45.Portable.csproj
@@ -28,16 +28,16 @@
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
     <DebugType>pdbonly</DebugType>
     <Optimize>true</Optimize>
-    <OutputPath>..\..\lib\portable45-net45+win8+wp8+wpa81\</OutputPath>
+    <OutputPath>..\..\lib\portable-net45+netcore45+wpa81+wp8\</OutputPath>
     <DefineConstants>TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="Api\ApiClient.net45.Portable.cs" />
-    <Compile Include="Api\Customers.net45.Portable.cs" />
-    <Compile Include="Api\Redemptions.net45.Portable.cs" />
-    <Compile Include="Api\Vouchers.net45.Portable.cs" />
+    <Compile Include="Api\Customers.net45.cs" />
+    <Compile Include="Api\Redemptions.net45.cs" />
+    <Compile Include="Api\Vouchers.net45.cs" />
     <Compile Include="Attributes\JsonEnumValueAttribute.cs" />
     <Compile Include="Constants.cs" />
     <Compile Include="Constants.net45.Portable.cs" />
@@ -98,12 +98,8 @@
       <Private>True</Private>
     </Reference>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
-  <Import Project="..\packages\Microsoft.Bcl.Build.1.0.14\tools\Microsoft.Bcl.Build.targets" Condition="Exists('..\packages\Microsoft.Bcl.Build.1.0.14\tools\Microsoft.Bcl.Build.targets')" />
-  <Target Name="EnsureBclBuildImported" BeforeTargets="BeforeBuild" Condition="'$(BclBuildImported)' == ''">
-    <Error Condition="!Exists('..\packages\Microsoft.Bcl.Build.1.0.14\tools\Microsoft.Bcl.Build.targets')" Text="This project references NuGet package(s) that are missing on this computer. Enable NuGet Package Restore to download them.  For more information, see http://go.microsoft.com/fwlink/?LinkID=317567." HelpKeyword="BCLBUILD2001" />
-    <Error Condition="Exists('..\packages\Microsoft.Bcl.Build.1.0.14\tools\Microsoft.Bcl.Build.targets')" Text="The build restored NuGet packages. Build the project again to include these packages in the build. For more information, see http://go.microsoft.com/fwlink/?LinkID=317568." HelpKeyword="BCLBUILD2002" />
-  </Target>
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
   <Target Name="BeforeBuild">
