+++ src/Voucherify.Client/Voucherify.Client.net45.csproj
@@ -32,10 +32,10 @@
     </DocumentationFile>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="Api\ApiClient.cs" />
-    <Compile Include="Api\Customers.cs" />
-    <Compile Include="Api\Redemptions.cs" />
-    <Compile Include="Api\Vouchers.cs" />
+    <Compile Include="Api\ApiClient.net45.cs" />
+    <Compile Include="Api\Customers.net45.cs" />
+    <Compile Include="Api\Redemptions.net45.cs" />
+    <Compile Include="Api\Vouchers.net45.cs" />
     <Compile Include="Attributes\JsonEnumValueAttribute.cs" />
     <Compile Include="Constants.cs" />
     <Compile Include="Constants.net45.cs" />
@@ -80,16 +80,8 @@
       <HintPath>..\packages\Newtonsoft.Json.9.0.1\lib\net45\Newtonsoft.Json.dll</HintPath>
       <Private>True</Private>
     </Reference>
-    <Reference Include="RestSharp, Version=105.2.3.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\RestSharp.105.2.3\lib\net45\RestSharp.dll</HintPath>
-      <Private>True</Private>
-    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Net.Http" />
-    <Reference Include="RSG.Promise, Version=1.2.0.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\RSG.Promise.1.2.0.0\lib\net35\RSG.Promise.dll</HintPath>
-      <Private>True</Private>
-    </Reference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
