+++ src/Voucherify.Client/packages.config
@@ -1,9 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Microsoft.Bcl" version="1.1.10" targetFramework="portable45-net45+win8+wp8+wpa81" />
-  <package id="Microsoft.Bcl.Build" version="1.0.14" targetFramework="portable45-net45+win8+wp8+wpa81" />
-  <package id="Microsoft.Net.Http" version="2.2.29" targetFramework="portable45-net45+win8+wp8+wpa81" />
+  <package id="Microsoft.Net.Http" version="2.2.29" />
   <package id="Newtonsoft.Json" version="9.0.1" />
-  <package id="RestSharp" version="105.2.3" />
-  <package id="RSG.Promise" version="1.2.0.0" />
 </packages>
